requirements.txt
python-dotenv
pandas
pytz
MetaTrader5
Pillow
discord.py
requests>=2.31.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
selenium>=4.25.0
webdriver-manager>=4.0.2
psutil
matplotlib
python-dateutil





# Optional - pro budoucí rozšíření
# plotly>=5.18.0  # Pro grafy
# matplotlib>=3.8.0  # Pro vizualizace
# numpy>=1.24.0  # Pro pokročilé výpočty